+++ packages/terra-content-container/tests/nightwatch/content-container-spec.js
@@ -1,19 +1,10 @@
-/* eslint import/no-extraneous-dependencies: ["error", {"devDependencies": true}] */
 /* eslint-disable no-unused-expressions */
 // eslint-disable-next-line import/no-extraneous-dependencies
-const screenshot = require('terra-toolkit').screenshot;
+const resizeTo = require('terra-toolkit/lib/nightwatch/responsive-helpers').resizeTo;
 
-module.exports = {
-  before: (browser, done) => {
-    browser.resizeWindow(browser.globals.width, browser.globals.height, done);
-  },
-
-  afterEach: (browser, done) => {
-    screenshot(browser, '#content-container', done);
-  },
-
+module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous'], {
   'Displays a content container with default props': (browser) => {
-    browser.url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/content-container-tests/default`);
+    browser.url(`${browser.launchUrl}/#/tests/content-container-tests/default`);
     browser.expect.element('#content-container').to.be.present;
     browser.expect.element('#content-container').to.have.attribute('class').which.not.contains('fill');
     browser.expect.element('#content-container div:nth-child(1)').text.to.contain('header');
@@ -22,7 +13,7 @@ module.exports = {
 
   'Displays a content container with fill': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/content-container-tests/fill`)
+      .url(`${browser.launchUrl}/#/tests/content-container-tests/fill`)
       .expect.element('#content-container').to.have.attribute('class').which.contains('fill');
   },
-};
+});