+++ packages/terra-table/tests/nightwatch/table-spec.js
@@ -1,25 +1,16 @@
-/* eslint import/no-extraneous-dependencies: ["error", {"devDependencies": true}] */
 // eslint-disable-next-line import/no-extraneous-dependencies
-const screenshot = require('terra-toolkit').screenshot;
+const resizeTo = require('terra-toolkit/lib/nightwatch/responsive-helpers').resizeTo;
 
-module.exports = {
-  before: (browser, done) => {
-    browser.resizeWindow(browser.globals.width, browser.globals.height, done);
-  },
-
-  afterEach: (browser, done) => {
-    screenshot(browser, 'terra-table', done);
-  },
-
+module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous'], {
   'Displays a default table with the provided text': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/table-tests/default`)
+      .url(`${browser.launchUrl}/#/tests/table-tests/default`)
       .assert.elementPresent('#Table')
       .assert.elementPresent('[class*="striped"]');
   },
 
   'Displays a default table and verify the provided text': (browser) => {
-    browser.url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/table-tests/default`);
+    browser.url(`${browser.launchUrl}/#/tests/table-tests/default`);
     browser.assert.containsText('th:nth-child(1)', 'Name');
     browser.expect.element('th:nth-child(1)').to.have.attribute('class').which.contains('min-width-small');
 
@@ -44,38 +35,38 @@ module.exports = {
 
   'Displays a table without the stripes': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/table-tests/no-striped`)
+      .url(`${browser.launchUrl}/#/tests/table-tests/no-striped`)
       .assert.elementPresent('#Table')
       .assert.elementNotPresent('[class*="striped"]');
   },
 
   'Displays a table without padding': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/table-tests/no-padding`)
+      .url(`${browser.launchUrl}/#/tests/table-tests/no-padding`)
       .assert.elementPresent('#Table')
       .assert.elementNotPresent('[class*="padded"]');
   },
 
   'Display a sorting indicator': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/table-tests/table-sort-indicator`)
+      .url(`${browser.launchUrl}/#/tests/table-tests/table-sort-indicator`)
       .assert.attributeContains('th:nth-child(2)', 'data-sort', 'asc');
   },
 
   'Display a table with selected rows': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/table-tests/table-highlighted-rows`)
+      .url(`${browser.launchUrl}/#/tests/table-tests/table-highlighted-rows`)
       .assert.elementPresent('tr[class*="is-selected"]');
   },
 
   'Displays a table with selectable rows': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/table-tests/selectable-table`)
+      .url(`${browser.launchUrl}/#/tests/table-tests/selectable-table`)
       .assert.elementPresent('tr[class*="is-selectable"]');
   },
 
   'Display a selectable table and highlights the selected row upon clicking': (browser) => {
-    browser.url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/table-tests/selectable-table`);
+    browser.url(`${browser.launchUrl}/#/tests/table-tests/selectable-table`);
     browser.click('[class*="row"]:nth-child(1)');
     browser.expect.element('[class*="row"]:nth-child(1)').to.have.attribute('class').which.contains('is-selected');
     browser.expect.element('[class*="row"]:nth-child(2)').to.have.attribute('class').which.not.contains('is-selected');
@@ -93,7 +84,7 @@ module.exports = {
   },
 
   'Toggle aria-selected on seletable rows': (browser) => {
-    browser.url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/table-tests/selectable-table`);
+    browser.url(`${browser.launchUrl}/#/tests/table-tests/selectable-table`);
     browser.click('[class*="row"]:nth-child(1)');
     browser.expect.element('[class*="row"]:nth-child(1)').to.have.attribute('aria-selected').which.contains('true');
     browser.expect.element('[class*="row"]:nth-child(2)').to.have.attribute('aria-selected').which.contains('false');
@@ -111,7 +102,7 @@ module.exports = {
   },
 
   'Display a selectable table and highlights the selected row upon enter': (browser) => {
-    browser.url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/table-tests/selectable-table`);
+    browser.url(`${browser.launchUrl}/#/tests/table-tests/selectable-table`);
 
     browser.sendKeys('[class*="row"]:nth-child(1)', browser.Keys.ENTER);
     browser.expect.element('[class*="row"]:nth-child(1)').to.have.attribute('class').which.contains('is-selected');
@@ -130,7 +121,7 @@ module.exports = {
   },
 
   'Display a selectable table and highlights the selected row upon space keydown': (browser) => {
-    browser.url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/table-tests/selectable-table`);
+    browser.url(`${browser.launchUrl}/#/tests/table-tests/selectable-table`);
 
     browser.sendKeys('[class*="row"]:nth-child(1)', browser.Keys.SPACE);
     browser.expect.element('[class*="row"]:nth-child(1)').to.have.attribute('class').which.contains('is-selected');
@@ -149,7 +140,7 @@ module.exports = {
   },
 
   'Display a selectable table with a subheader': (browser) => {
-    browser.url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/table-tests/table-selectable-subheaders`);
+    browser.url(`${browser.launchUrl}/#/tests/table-tests/table-selectable-subheaders`);
     browser.expect.element('tbody tr:nth-child(1)').to.have.attribute('data-terra-table-subheader-row');
     browser.assert.containsText('tbody tr:nth-child(1) td:nth-child(1)', 'Single');
     browser.expect.element('tbody tr:nth-child(4)').to.have.attribute('data-terra-table-subheader-row');
@@ -157,7 +148,7 @@ module.exports = {
   },
 
   'Display a non-selectable table with a subheader': (browser) => {
-    browser.url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/table-tests/table-subheaders`);
+    browser.url(`${browser.launchUrl}/#/tests/table-tests/table-subheaders`);
     browser.expect.element('tbody tr:nth-child(1)').to.have.attribute('data-terra-table-subheader-row');
     browser.assert.containsText('tbody tr:nth-child(1) td:nth-child(1)', 'Single');
     browser.expect.element('tbody tr:nth-child(4)').to.have.attribute('data-terra-table-subheader-row');
@@ -165,7 +156,7 @@ module.exports = {
   },
 
   'Triggers onChange for selectable table upon clicking a row': (browser) => {
-    browser.url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/table-tests/selectable-table-onchange`);
+    browser.url(`${browser.launchUrl}/#/tests/table-tests/selectable-table-onchange`);
 
     browser.click('[class*="row"]:nth-child(1)');
     browser.expect.element('[class*="row"]:nth-child(1)').to.have.attribute('class').which.contains('is-selected');
@@ -187,33 +178,29 @@ module.exports = {
   },
 
   'Displays a table with only one row': (browser) => {
-    browser
-    .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/table-tests/single-row-table`);
+    browser.url(`${browser.launchUrl}/#/tests/table-tests/single-row-table`);
     browser.assert.elementPresent('[class*="row"]:nth-child(1)');
     browser.assert.elementNotPresent('[class*="row"]:nth-child(2)');
   },
 
   'Displays a table with no rows': (browser) => {
-    browser
-    .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/table-tests/table-no-rows`);
+    browser.url(`${browser.launchUrl}/#/tests/table-tests/table-no-rows`);
     browser.assert.elementNotPresent('[class*="row"]:nth-child(1)');
   },
 
   'Displays a selectable table with only one row': (browser) => {
-    browser
-    .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/table-tests/selectable-table-single-row`);
+    browser.url(`${browser.launchUrl}/#/tests/table-tests/selectable-table-single-row`);
     browser.assert.elementPresent('[class*="row"]:nth-child(1)');
     browser.assert.elementNotPresent('[class*="row"]:nth-child(2)');
   },
 
   'Displays a selectable table with no rows': (browser) => {
-    browser
-    .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/table-tests/selectable-table-no-rows`);
+    browser.url(`${browser.launchUrl}/#/tests/table-tests/selectable-table-no-rows`);
     browser.assert.elementNotPresent('[class*="row"]:nth-child(1)');
   },
 
   'Maintains custom classNames when provided': (browser) => {
-    browser.url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/table-tests/table-selectable-subheaders`);
+    browser.url(`${browser.launchUrl}/#/tests/table-tests/table-selectable-subheaders`);
     browser.assert.elementPresent('.Table-Custom');
     browser.assert.elementPresent('.Header-Custom');
     browser.assert.elementPresent('.HeaderCell-Custom');
@@ -221,4 +208,4 @@ module.exports = {
     browser.assert.elementPresent('.SingleSelectableRows-Custom');
     browser.assert.elementPresent('.Subheader-Custom');
   },
-};
+});