+++ packages/terra-badge/CHANGELOG.md
@@ -3,13 +3,18 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Update nightwatch tests and test scripts.
+
 ### Added
 * Added cerner-mock-theme styles
+
 
 1.5.0 - (August 31, 2017)
 ------------------
 ### Added
 * Added cerner-mock-theme styles
+
 
 1.4.0 - (August 15, 2017)
 ------------------