+++ packages/terra-i18n/CHANGELOG.md
@@ -7,6 +7,9 @@ Unreleased
 1.4.0 - (August 31, 2017)
 ------------------
 ### Changed
+
+* Update nightwatch tests and test scripts.
+
 * Improved terra-i18n example in README.md
 
 1.3.0 - (August 8, 2017)