+++ packages/terra-image/CHANGELOG.md
@@ -9,6 +9,9 @@ Unreleased
 ### Added
 * Added default border, margin, and padding reset to img
 
+### Changed
+* Update nightwatch tests and test scripts.
+
 1.4.0 - (August 15, 2017)
 ------------------
 ### Changed